refactor: replace session updates endpoint with last-block-id
Replace the complex long-polling `/updates` endpoint with a simpler `/last-block-id` endpoint that returns only the most recent block ID. Add `turnDetail` function to fetch a single turn with all its blocks.

- Rename `/api/projects/:projectId/sessions/:sessionId/updates` to `/api/projects/:projectId/sessions/:sessionId/last-block-id`
- Simplify response from full session data to `{ lastBlockId: string | null }`
- Remove long-polling logic and state tracking
- Add `turnDetail` function in workspace signals to fetch turn with blocks
- Update all contracts, tests, and MSW handlers
- Clean up unused schemas and type exports
