+++ turbo/apps/web/app/api/projects/[projectId]/sessions/[sessionId]/last-block-id/route.test.ts
@@ -1,0 +1,245 @@
+import { describe, it, expect, beforeEach, vi, afterEach } from "vitest";
+import "../../../../../../../src/test/setup";
+import { NextRequest } from "next/server";
+import { GET } from "./route";
+import { initServices } from "../../../../../../../src/lib/init-services";
+import { PROJECTS_TBL } from "../../../../../../../src/db/schema/projects";
+import {
+  SESSIONS_TBL,
+  TURNS_TBL,
+  BLOCKS_TBL,
+} from "../../../../../../../src/db/schema/sessions";
+import { eq } from "drizzle-orm";
+import * as Y from "yjs";
+
+// Mock Clerk authentication
+vi.mock("@clerk/nextjs/server", () => ({
+  auth: vi.fn(),
+}));
+
+import { auth } from "@clerk/nextjs/server";
+const mockAuth = vi.mocked(auth);
+
+describe("/api/projects/:projectId/sessions/:sessionId/last-block-id", () => {
+  const projectId = `lastblock-${Date.now()}`;
+  const sessionId = `sess_lastblock_${Date.now()}`;
+  const userId = `test-user-lastblock-${Date.now()}-${process.pid}`;
+  let createdTurnIds: string[] = [];
+  let createdBlockIds: string[] = [];
+
+  beforeEach(async () => {
+    vi.clearAllMocks();
+    // Mock successful authentication by default
+    mockAuth.mockResolvedValue({ userId } as Awaited<ReturnType<typeof auth>>);
+
+    // Initialize services
+    initServices();
+
+    // Clean up any existing test data
+    await globalThis.services.db
+      .delete(PROJECTS_TBL)
+      .where(eq(PROJECTS_TBL.id, projectId));
+
+    // Create test project
+    const ydoc = new Y.Doc();
+    const state = Y.encodeStateAsUpdate(ydoc);
+    const base64Data = Buffer.from(state).toString("base64");
+
+    await globalThis.services.db.insert(PROJECTS_TBL).values({
+      id: projectId,
+      userId,
+      ydocData: base64Data,
+      version: 0,
+    });
+
+    // Create test session
+    await globalThis.services.db.insert(SESSIONS_TBL).values({
+      id: sessionId,
+      projectId,
+      title: "Test Session for Last Block ID",
+    });
+
+    createdTurnIds = [];
+    createdBlockIds = [];
+  });
+
+  afterEach(async () => {
+    // Clean up blocks
+    for (const blockId of createdBlockIds) {
+      await globalThis.services.db
+        .delete(BLOCKS_TBL)
+        .where(eq(BLOCKS_TBL.id, blockId));
+    }
+
+    // Clean up turns
+    for (const turnId of createdTurnIds) {
+      await globalThis.services.db
+        .delete(TURNS_TBL)
+        .where(eq(TURNS_TBL.id, turnId));
+    }
+
+    // Clean up session
+    await globalThis.services.db
+      .delete(SESSIONS_TBL)
+      .where(eq(SESSIONS_TBL.id, sessionId));
+
+    // Clean up project
+    await globalThis.services.db
+      .delete(PROJECTS_TBL)
+      .where(eq(PROJECTS_TBL.id, projectId));
+  });
+
+  describe("GET /api/projects/:projectId/sessions/:sessionId/last-block-id", () => {
+    it("should return 401 when not authenticated", async () => {
+      mockAuth.mockResolvedValueOnce({ userId: null } as Awaited<
+        ReturnType<typeof auth>
+      >);
+
+      const request = new NextRequest("http://localhost:3000");
+      const context = { params: Promise.resolve({ projectId, sessionId }) };
+
+      const response = await GET(request, context);
+
+      expect(response.status).toBe(401);
+      const data = await response.json();
+      expect(data).toHaveProperty("error", "unauthorized");
+    });
+
+    it("should return 404 when project doesn't exist", async () => {
+      const request = new NextRequest("http://localhost:3000");
+      const context = {
+        params: Promise.resolve({
+          projectId: "non-existent",
+          sessionId,
+        }),
+      };
+
+      const response = await GET(request, context);
+
+      expect(response.status).toBe(404);
+      const data = await response.json();
+      expect(data).toHaveProperty("error", "project_not_found");
+    });
+
+    it("should return 404 when session doesn't exist", async () => {
+      const request = new NextRequest("http://localhost:3000");
+      const context = {
+        params: Promise.resolve({
+          projectId,
+          sessionId: "non-existent",
+        }),
+      };
+
+      const response = await GET(request, context);
+
+      expect(response.status).toBe(404);
+      const data = await response.json();
+      expect(data).toHaveProperty("error", "session_not_found");
+    });
+
+    it("should return null when session has no blocks", async () => {
+      const request = new NextRequest("http://localhost:3000");
+      const context = { params: Promise.resolve({ projectId, sessionId }) };
+
+      const response = await GET(request, context);
+
+      expect(response.status).toBe(200);
+      const data = await response.json();
+      expect(data).toEqual({ lastBlockId: null });
+    });
+
+    it("should return the last block ID when session has one block", async () => {
+      const [turn] = await globalThis.services.db
+        .insert(TURNS_TBL)
+        .values({
+          id: `turn_${Date.now()}`,
+          sessionId,
+          userPrompt: "Question",
+          status: "completed",
+        })
+        .returning();
+
+      const [block] = await globalThis.services.db
+        .insert(BLOCKS_TBL)
+        .values({
+          id: `block_${Date.now()}`,
+          turnId: turn!.id,
+          type: "content",
+          content: { text: "Answer" },
+          sequenceNumber: 0,
+        })
+        .returning();
+
+      createdTurnIds.push(turn!.id);
+      createdBlockIds.push(block!.id);
+
+      const request = new NextRequest("http://localhost:3000");
+      const context = { params: Promise.resolve({ projectId, sessionId }) };
+
+      const response = await GET(request, context);
+
+      expect(response.status).toBe(200);
+      const data = await response.json();
+      expect(data).toEqual({ lastBlockId: block!.id });
+    });
+
+    it("should return the most recent block ID when session has multiple blocks", async () => {
+      const [turn1] = await globalThis.services.db
+        .insert(TURNS_TBL)
+        .values({
+          id: `turn_1_${Date.now()}`,
+          sessionId,
+          userPrompt: "Question 1",
+          status: "completed",
+        })
+        .returning();
+
+      const [block1] = await globalThis.services.db
+        .insert(BLOCKS_TBL)
+        .values({
+          id: `block_1_${Date.now()}`,
+          turnId: turn1!.id,
+          type: "content",
+          content: { text: "Answer 1" },
+          sequenceNumber: 0,
+        })
+        .returning();
+
+      // Wait a bit to ensure different timestamps
+      await new Promise((resolve) => setTimeout(resolve, 10));
+
+      const [turn2] = await globalThis.services.db
+        .insert(TURNS_TBL)
+        .values({
+          id: `turn_2_${Date.now()}`,
+          sessionId,
+          userPrompt: "Question 2",
+          status: "completed",
+        })
+        .returning();
+
+      const [block2] = await globalThis.services.db
+        .insert(BLOCKS_TBL)
+        .values({
+          id: `block_2_${Date.now()}`,
+          turnId: turn2!.id,
+          type: "content",
+          content: { text: "Answer 2" },
+          sequenceNumber: 0,
+        })
+        .returning();
+
+      createdTurnIds.push(turn1!.id, turn2!.id);
+      createdBlockIds.push(block1!.id, block2!.id);
+
+      const request = new NextRequest("http://localhost:3000");
+      const context = { params: Promise.resolve({ projectId, sessionId }) };
+
+      const response = await GET(request, context);
+
+      expect(response.status).toBe(200);
+      const data = await response.json();
+      expect(data).toEqual({ lastBlockId: block2!.id });
+    });
+  });
+});
