+++ turbo/apps/web/src/test/msw-handlers.ts
@@ -274,32 +274,13 @@ const projectsHandlers = [
     },
   ),
 
-  // GET /api/projects/:projectId/sessions/:sessionId/updates - Long polling updates
+  // GET /api/projects/:projectId/sessions/:sessionId/last-block-id - Get last block ID
   http.get(
-    "*/api/projects/:projectId/sessions/:sessionId/updates",
-    async ({ request }) => {
-      const url = new URL(request.url);
-      const timeout = url.searchParams.get("timeout");
-
-      // For immediate requests (refetch), return 204 immediately
-      if (timeout === "0") {
-        return new HttpResponse(null, { status: 204 });
-      } else {
-        // For long polling requests, simulate waiting but respond to abort signals
-        return new Promise((resolve, reject) => {
-          const timeoutId = setTimeout(() => {
-            resolve(new HttpResponse(null, { status: 204 }));
-          }, 200); // Short delay for tests instead of 30 seconds
-
-          // Listen for abort signal
-          if (request.signal) {
-            request.signal.addEventListener("abort", () => {
-              clearTimeout(timeoutId);
-              reject(new Error("Aborted"));
-            });
-          }
-        });
-      }
+    "*/api/projects/:projectId/sessions/:sessionId/last-block-id",
+    () => {
+      return HttpResponse.json({
+        lastBlockId: null,
+      });
     },
   ),
